Add minor changes

+++ src/main/java/com/bp/RUIAN/repositories/AddressRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
 
 @EnableElasticsearchRepositories
-public interface AddressRepository extends ElasticsearchRepository<Address, Long> {
+public interface AddressRepository extends ElasticsearchRepository<Address, Integer> {
 
 }
